requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-channels==4.0.0
channels-redis==4.1.0
django-bootstrap5==23.3
django-spectacular==0.26.5
pyodbc==5.0.1
redis==5.0.1
Pillow==10.1.0
pandas
chardet
